feat(form): adjusted form page layout
