+++ src/components/index.ts
@@ -1,3 +1,5 @@
 import FormInput from "./FormInput";
-
-export { FormInput };
+import FormSelect from "./FormSelect";
+import Sidebar from "./Sidebar";
+import Navbar from "./Navbar";
+export { FormInput, FormSelect, Sidebar, Navbar };
